Fix infinite loop on parent pom.xml with same artifactId
Root cause of the issue:

Trivy needs to get the parent POM for
every dependency it analyzes, and it starts by
checking ../pom.xml for the parent before making
an HTTP call to Maven central. Most of the time
../pom.xml will NOT be the parent POM for a
dependency. Trivy detects whether ../pom.xml is the
parent by comparing the parent and dependency
artifactId (to fix this same infinite loop bug)
but in some cases, the artifactIds but not the
groupIds match, causing a different type of
infinite loop.